HD-1598. File Write errors
Datanode side accounting of dfsUsed, numBlocks and reservedBytes is a joke. I propose internally we ignore reservations and wait for it to be done properly in the community. If the improper accounting of dfsUsed and numBlocks causes issues, we should look to ignore those too in the future. Not doing that just yet. Here are some of the issues reported already (and many more exist which have not been reported)

https://issues.apache.org/jira/browse/HDFS-9530
https://issues.apache.org/jira/browse/HDFS-6489
https://issues.apache.org/jira/browse/HDFS-10271
(cherry picked from commit b763f7f7175cab1303b654a9852ae0fbe4b9fb7d)